// typescript-sdk/src/server/types/tracer.ts
type JSONSerializable =
  | string
  | number
  | boolean
  | null
  | Record<string, any>
  | any[];

interface TypedValueJson {
  type: "json";
  value: JSONSerializable;
}

type Money = {
  currency: string;
  amount: number;
};

export interface GuardrailResult {
  status: "processed" | "skipped" | "error";
  passed: boolean;
  score?: number | null;
  details?: string | null;
  cost?: Money | null;
}

interface TypedValueGuardrailResult {
  type: "guardrail_result";
  value: GuardrailResult;
}
// ...
export type SpanTypes =
  | "span"
  | "llm"
  | "chain"
  | "tool"
  | "agent"
  | "rag"
  | "guardrail"
  | "unknown";
// ...
export interface RAGChunk {
  document_id?: string | null;
  chunk_id?: string | null;
  content: string | Record<string, any> | any[];
}